split all modules into header+cpp
