+++ ue2/model_tlm_lt/src/memory_manager.cpp
@@ -1,0 +1,48 @@
+#include "memory_manager.h"
+
+mm::mm()
+    : free_list(0),
+      empties(0)
+#ifdef DEBUG
+      ,
+      count(0)
+#endif
+{
+}
+
+mm::gp_t* mm::allocate() {
+#ifdef DEBUG
+  fout << "----------------------------- Called allocate(), #trans = "
+       << ++count << endl;
+#endif
+  gp_t* ptr;
+
+  if (free_list) {
+    ptr = free_list->trans;
+    empties = free_list;
+    free_list = free_list->next;
+  } else {
+    ptr = new gp_t(this);
+  }
+  return ptr;
+}
+
+void mm::free(gp_t* trans) {
+#ifdef DEBUG
+  fout << "----------------------------- Called free(), #trans = " << --count
+       << endl;
+#endif
+
+  if (!empties) {
+    empties = new access;
+    empties->next = free_list;
+    empties->prev = 0;
+
+    if (free_list) {
+      free_list->prev = empties;
+    }
+  }
+  free_list = empties;
+  free_list->trans = trans;
+  empties = free_list->prev;
+}
