+++ ue2/model_tlm_lt/src/memory_manager.h
@@ -1,17 +1,12 @@
-#ifndef MEMORY_MANAGER_H_
-#define MEMORY_MANAGER_H_
+#ifndef _MEMORY_MANAGER_H_
+#define _MEMORY_MANAGER_H_
 
 #define SC_INCLUDE_DYNAMIC_PROCESSES
 
 #include <systemc.h>
-// using namespace sc_core;
-// using namespace sc_dt;
-// using namespace std;
-
 #include <tlm.h>
 #include <tlm_utils/peq_with_cb_and_phase.h>
 #include <tlm_utils/simple_initiator_socket.h>
-#include <tlm_utils/simple_target_socket.h>
 
 static ofstream fout("memory_manager.log");
 
@@ -22,15 +17,7 @@ class mm : public tlm::tlm_mm_interface {
   typedef tlm::tlm_generic_payload gp_t;
 
  public:
-  mm()
-      : free_list(0),
-        empties(0)
-#ifdef DEBUG
-        ,
-        count(0)
-#endif
-  {
-  }
+  mm();
 
   gp_t* allocate();
   void free(gp_t* trans);
@@ -50,41 +37,4 @@ class mm : public tlm::tlm_mm_interface {
 #endif
 };
 
-mm::gp_t* mm::allocate() {
-#ifdef DEBUG
-  fout << "----------------------------- Called allocate(), #trans = "
-       << ++count << endl;
-#endif
-  gp_t* ptr;
-
-  if (free_list) {
-    ptr = free_list->trans;
-    empties = free_list;
-    free_list = free_list->next;
-  } else {
-    ptr = new gp_t(this);
-  }
-  return ptr;
-}
-
-void mm::free(gp_t* trans) {
-#ifdef DEBUG
-  fout << "----------------------------- Called free(), #trans = " << --count
-       << endl;
-#endif
-
-  if (!empties) {
-    empties = new access;
-    empties->next = free_list;
-    empties->prev = 0;
-
-    if (free_list) {
-      free_list->prev = empties;
-    }
-  }
-  free_list = empties;
-  free_list->trans = trans;
-  empties = free_list->prev;
-}
-
-#endif /* MEMORY_MANAGER_H_ */
+#endif /* _MEMORY_MANAGER_H_ */
